+++ src/Thebes.jl
@@ -3,13 +3,10 @@ module Thebes
 using Luxor
 using StaticArrays
 using Rotations
-using Requires
-#using LaTexStrings
+using PackageExtensionCompat
 
 function __init__()
-    @require MathTeXEngine = "0a4f8689-d25c-4efe-a92b-7142dfc1aa53" begin
-        @require LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f" include("latex.jl")
-    end
+    @require_extensions
 end
 
 include("Point3D.jl")