+++ test/runtests.jl
@@ -11,6 +11,7 @@ mktempdir() do tmpdir
 
         @info "...testing 3D point arithmetic"
         include("point3d-arithmetic.jl")
+
         @info "... running test-1"
         include("test-1.jl")
 
@@ -31,6 +32,9 @@ mktempdir() do tmpdir
 
         @info "... running test-7"
         include("test-7.jl")
+
+        @info "... running latex text"
+        include("latex-test.jl")
 
         @info "Write more tests...!"
     end